Check for presence of __builtin_mul_overflow
If that's not there, use a software implementation (which assumes
size_t).

Caveat: mallocr.c uses a macro `INTERNAL_SIZE_T`, which defaults
to `size_t` in case that macro is not defined. So in case that macro
is defined, an `#error` has been added to avoid subtle errors.

(Cherry-picked from branch 'compcert')